ensure correct model for gubernator tests
The gubernator tests rely on "juju switch" being in the right
model before processing other juju commands. This is fragile
because an admin or other job might switch the controller during
a test run.

Ensure all juju commands for gubernator tests specify the
'controller:model' option.

Note, this also involves a change in the e2e-gubernator-* jobs
in jenkins to export a new CONTROLLER envar before the test begins.