Fixing easyrsa test
Changing easyrsa test to no longer scrape the disk since the client cert isn't left on disk anymore.
It looks at the relation data now. We could do this for the ca certificate, but we don't have the
ca key in the relation data, so I'm leaving the ca test for now.