Deploy hacluster only if one is not deployed yet
Right now test_hacluster is just blindly deploying an hacluster. If
there is one already deployed, it should use that instead. Currently this just
searches through application names in the Juju model and checks for the string
'hacluster'. It would be nice if there was a more robust way to check,
but this is at least a step in right direction.