Limit attempt in test_anon_auth_disabled and provide more debugging info
Turns out that #610 didn't fix the kubelet connection issue and just
ended up timing the job out after 3 hours. This sets the upper limit to
just over 2 minutes and if it still can't connect, it will attempt to
collect and report some additional debugging info to try to track down
why this is happening.