Fix handling of paths for bundle push (#640)
The bundle build process puts the results into subdirectories of the
cwd, but the previous change tried to make the push be relative to the
cloned bundle repo dir.  Instead, this reverts it back to using the cwd
subdirectory and makes skip-build bundles copy the bundle info to a cwd
subdir to match that pattern.