Improve cloud filter and add Octavia test
The filter for running tests on only specific clouds was driven by the
`--cloud` param rather than the actual cloud the test is being run on,
leading to tests being skipped if the cloud param wasn't given
(defaulting to "aws").

This also adds a test for Octavia when running on OpenStack.