Drop openstack_integrator and improve skip_by_cloud (#732)
The openstack_integrator fixture was running even on unsupported clouds
due to scope conflicts with skip_by_cloud. Added to that, the fixture
was almost guaranteed to cause the Octavia test to fail anyway due to
not knowing the proper values to configure for the subnet-id and
floating-network-id configs, so we really need to ensure that the
integrator is pre-deployed and configured properly.

Additionally, the skip_by_cloud fixture wasn't doing any validation of
the clouds and the previous change to use the actual cloud type meant
that the tests marked with "aws" were being skipped incorrectly.