Update tigera_aws.py (#1616)
* Update tigera_aws.py

Add .replace call to REGION. This is coupled to the cloud provider but I believe it's OK since the file is called _aws.py. Also, I did not change JUJU_CLOUD because aws/us-east-2 is valid for a JUJU_CLOUD.

I changed the REGION variable because aws/us-east-2 is not a valid region.